LocalProjections

LocalProjections

LocalProjections.LocalProjection

LocalProjections.F!

LocalProjections.FΔ!

LocalProjections.L!

LocalProjections.LΔ!

LocalProjections.coef

LocalProjections.coef_se

LocalProjections.makedepvars

LocalProjections.runregs

LocalProjections.se

LocalProjections.Δ

LocalProjections.LocalProjection — Type

Struct with the following fields

df

horizons

models

transformfunc

grwthoffset

formula

regstyle

ynames

LocalProjections.F! — Function

F!(df, v, horizons=1, fun=identity)

Forward operator. Applies function fun to the column v of df and then makes the leads specified by horizons and adds them to df as the column F(i)(v). Modifies df. Return the list of new columns created.

LocalProjections.FΔ! — Function

F!(df, v, horizons=1, fun=identity, baseperiod=0, relativebase=false)

Cumlative forward difference operator Applies function fun to the column v of df and then calculates the growth of this column, saving the results to a new columns in df as the column FΔ(i)(v). If x = map(fun, df[!, v]), df.FΔ(i)(v) = lead(x, horizons) - lead(x, baseperiod) or x = map(fun, df[!, v]), df.FΔ(i)(v) = lead(x, horizons) - lead(x, horizon + baseperiod) if relativebase == true Modifies df. Return the list of new columns created.

LocalProjections.L! — Function

L!(df, v, horizons=1, fun=identity)

Lag operator. Applies function fun to the column v of df and then makes the laggs specified by horizons and adds them to df as the column L(i)(v). Modifies df. Return the list of new columns created.

LocalProjections.LΔ! — Function

LΔ!(df, v, horizons=1, fun=identity, baseperiod=0, relativebase=false)

Cumlative lag difference operator

Applies function fun to the column v of df and then calculates the lagged growth of this column, saving the results to a new columns in df as the column LΔ(i)(v). If x = map(fun, df[!, v]), df.FΔ(i)(v) = lag(x, baseperiod) - lag(x, horizons) or x = map(fun, df[!, v]), df.FΔ(i)(v) = lag(x, i+baseperiod) - lag(x, horizons) if relativebase == true Modifies df. Return the list of new columns created.

LocalProjections.coef — Method

coef(lp::LocalProjection, xname::Symbol)

Returns a vector of coefficients for the variable xname for each regression. coef(lp, xname)[i] corresponds to the coefficient for the horizon lp.horizons[i].

LocalProjections.coef_se — Method

coef_se(lp::LocalProjection, xname::Symbol)

Returns a dataframe with the columns [horizons, β, se] aligning the estimated coefficient, standard error and horizon that these correspond to. Constructed from DataFrame(horizons = lp.horizons, β = coef(lp, xname), se = se(lp, xname))

LocalProjections.makedepvars — Method

makedepvars(lp::LocalProjection)

Takes the lhs variable from lp.formula and makes the cumlative growth variables for each horizon specified by horizons for the y-variable. This updates the dataframe stored in lp. Also updates the ynames array with the name of the created variables

LocalProjections.runregs — Method

runregs(lp::LocalProjection)

Estimates the regression for each horizon using FixedEffectModels.reg(lp.df, Term(lp.ynames[index]) ~ lp.formula.rhs) and store the estimated models in lp.models. The estimations are done lazely, i.e. each regression starts from scratch without saving intermediate computations common across each regression

LocalProjections.se — Method

se(lp::LocalProjection, xname::Symbol)

Returns a vector of standard errors for each coefficient for the variable xname for each regression. se(lp, xname)[i] corresponds to the se for the horizon lp.horizons[i].

LocalProjections.Δ — Function

Δ(x, horizons=1, fun=identity)

Single period growth function map(fun, x) .- lag(map(fun, x), horizons)