LocalProjections

LocalProjections.LocalProjection

LocalProjections.F!

LocalProjections.FΔ!

LocalProjections.L!

LocalProjections.LΔ!

LocalProjections.Δ

LocalProjections.LocalProjection — Type

Struct with the following fields

df

horizons

models

transformfunc

grwthoffset

formula

regstyle

ynames

LocalProjections.F! — Function

F!(df, v, horizons=1, fun=identity)

Forward operator. Applies function fun to the column v of df and then makes the leads specified by horizons and adds them to df as the column F(i)(v). Modifies df. Return the list of new columns created.

LocalProjections.FΔ! — Function

F!(df, v, horizons=1, fun=identity, baseperiod=0, relativebase=false)

Cumlative forward difference operator Applies function fun to the column v of df and then calculates the growth of this column, saving the results to a new columns in df as the column FΔ(i)(v). If x = map(fun, df[!, v]), df.FΔ(i)(v) = lead(x, horizons) - lead(x, baseperiod) or x = map(fun, df[!, v]), df.FΔ(i)(v) = lead(x, horizons) - lead(x, horizon + baseperiod) if relativebase == true Modifies df. Return the list of new columns created.

LocalProjections.L! — Function

L!(df, v, horizons=1, fun=identity)

Lag operator. Applies function fun to the column v of df and then makes the laggs specified by horizons and adds them to df as the column L(i)(v). Modifies df. Return the list of new columns created.

LocalProjections.LΔ! — Function

LΔ!(df, v, horizons=1, fun=identity, baseperiod=0, relativebase=false)

Cumlative lag difference operator

Applies function fun to the column v of df and then calculates the lagged growth of this column, saving the results to a new columns in df as the column LΔ(i)(v). If x = map(fun, df[!, v]), df.FΔ(i)(v) = lag(x, baseperiod) - lag(x, horizons) or x = map(fun, df[!, v]), df.FΔ(i)(v) = lag(x, i+baseperiod) - lag(x, horizons) if relativebase == true Modifies df. Return the list of new columns created.

LocalProjections.Δ — Function

Δ(x, horizons=1, fun=identity)

Single period growth function map(fun, x) .- lag(map(fun, x), horizons)